requirements.txt
astropy==5.2.2
asttokens==2.2.1
backcall==0.2.0
blosc2==2.0.0
bokeh==2.4.3
cdasws==1.7.42
cdflib==0.4.9
certifi==2023.5.7
cftime==1.6.2
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.39.4
geopack==1.0.10
hapiclient==0.2.5
idna==3.4
ipykernel==6.23.0
ipympl==0.9.3
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
netCDF4==1.6.3
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyerfa==2.0.0.3
Pygments==2.15.1
pyparsing==3.0.9
pyspedas==1.4.36
python-dateutil==2.8.2
pytplot-mpl-temp==2.1.37
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tables==3.8.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
viresclient==0.11.1
wcwidth==0.2.6
widgetsnbextension==4.0.7
xarray==2023.4.2
